demo/requirements.txt
tensorflow
h5py
Pillow
numpy
scikit-learn
torchvision
torch
matplotlib
tqdm
tkinterdnd2
opencv-python